fix: routerId 자료형 변경

+++ src/main/java/in/koreatech/batch/domain/bus/dto/CityBusRouteInfoApiResponse.java
@@ -13,7 +13,7 @@ public record CityBusRouteInfoApiResponse(
     public record CityBusRouteInfo(
 
         @JsonProperty("ROUTE_ID")
-        Long routeId,
+        Integer routeId,
 
         @JsonProperty("ROUTE_NAME")
         String routeName,
